+++ feature-account-api/src/main/java/io/novafoundation/nova/feature_account_api/presenatation/mixin/addressInput/AddressInputMixin.kt
@@ -1,22 +1,17 @@
 package io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput
 
-import androidx.lifecycle.LiveData
-import io.novafoundation.nova.common.domain.ExtendedLoadingState
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.ExternalAccountResolver
 import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.inputSpec.AddressInputSpec
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 
-interface AddressInputMixin {
+interface AddressInputMixin : ExternalAccountResolver {
 
     suspend fun getInputSpec(): AddressInputSpec
 
     val inputFlow: MutableStateFlow<String>
 
     val state: Flow<AddressInputState>
-
-    val externalIdentifierEventLiveData: LiveData<AccountIdentifierProvider.Event>
-
-    val selectedExternalAccountFlow: Flow<ExtendedLoadingState<ExternalAccount?>>
 
     fun pasteClicked()
 
@@ -25,12 +20,6 @@ interface AddressInputMixin {
     fun scanClicked()
 
     fun myselfClicked()
-
-    fun selectedExternalAddressClicked()
-
-    fun loadExternalIdentifiers()
-
-    fun selectExternalAccount(it: ExternalAccount)
 
     suspend fun getAddress(): String
