--- feature-account-api/src/main/java/io/novafoundation/nova/feature_account_api/presenatation/mixin/addressInput/InputFlowProvider.kt
@@ -1,12 +1,0 @@
-package io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput
-
-import kotlinx.coroutines.flow.MutableStateFlow
-
-interface InputFlowProvider {
-
-    val inputFlow: MutableStateFlow<String>
-}
-
-class RealInputFlowProvider : InputFlowProvider {
-    override val inputFlow = MutableStateFlow<String>("")
-}
