+++ feature-account-api/src/main/java/io/novafoundation/nova/feature_account_api/presenatation/mixin/addressInput/externalAccount/providers/Web3NameIdentifierProvider.kt
@@ -1,17 +1,19 @@
-package io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput
 
-import androidx.lifecycle.LiveData
+package io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.providers
+
 import androidx.lifecycle.MutableLiveData
 import io.novafoundation.nova.common.domain.ExtendedLoadingState
 import io.novafoundation.nova.common.domain.emitLoaded
 import io.novafoundation.nova.common.domain.emitLoading
-import io.novafoundation.nova.common.domain.loadedNothing
 import io.novafoundation.nova.common.presentation.toShortAddressFormat
 import io.novafoundation.nova.common.resources.ResourceManager
 import io.novafoundation.nova.feature_account_api.R
-import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.AccountIdentifierProvider.Event.ErrorEvent
-import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.AccountIdentifierProvider.Event.ShowBottomSheetEvent
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.AccountIdentifierProvider
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.AccountIdentifierProvider.Event.ErrorEvent
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.AccountIdentifierProvider.Event.ShowBottomSheetEvent
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.externalAccount.ExternalAccount
 import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.inputSpec.AddressInputSpecProvider
+import io.novafoundation.nova.feature_account_api.presenatation.mixin.addressInput.toIdenticonState
 import io.novafoundation.nova.runtime.multiNetwork.ChainWithAsset
 import io.novafoundation.nova.web3names.domain.exceptions.Web3NamesException
 import io.novafoundation.nova.web3names.domain.models.Web3NameAccount
@@ -23,59 +25,16 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combineTransform
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-
-interface AccountIdentifierProvider {
-
-    val selectedExternalAccountFlow: Flow<ExtendedLoadingState<ExternalAccount?>>
-
-    val eventsLiveData: LiveData<Event>
-
-    fun selectExternalAccount(account: ExternalAccount?)
-
-    fun isIdentifierValid(raw: String): Boolean
-
-    fun loadExternalAccounts(raw: String)
-
-    sealed interface Event {
-
-        class ShowBottomSheetEvent(
-            val identifier: String,
-            val chainName: String,
-            val externalAccounts: List<ExternalAccount>,
-            val selectedAccount: ExternalAccount?
-        ) : Event
-
-        class ErrorEvent(val exception: Throwable) : Event
-    }
-}
-
-class EmptyAccountIdentifierProvider : AccountIdentifierProvider {
-
-    override val selectedExternalAccountFlow: Flow<ExtendedLoadingState<ExternalAccount?>> = flowOf(loadedNothing())
-
-    override val eventsLiveData: LiveData<AccountIdentifierProvider.Event> = MutableLiveData()
-
-    override fun selectExternalAccount(account: ExternalAccount?) {
-        // empty implementation
-    }
-
-    override fun loadExternalAccounts(raw: String) {
-        // empty implementation
-    }
-
-    override fun isIdentifierValid(raw: String) = false
-}
 
 class Web3NameIdentifierProvider(
     private val web3NameInteractor: Web3NamesInteractor,
     private val destinationChain: Flow<ChainWithAsset>,
     private val addressInputSpecProvider: AddressInputSpecProvider,
     private val coroutineScope: CoroutineScope,
-    private val inputFlowProvider: InputFlowProvider,
+    private val input: Flow<String>,
     private val resourceManager: ResourceManager
 ) : AccountIdentifierProvider,
     CoroutineScope by coroutineScope {
@@ -98,7 +57,7 @@ class Web3NameIdentifierProvider(
     override val eventsLiveData = MutableLiveData<AccountIdentifierProvider.Event>()
 
     init {
-        inputFlowProvider.inputFlow.onEach {
+        input.onEach {
             _selectedExternalAccountFlow.value = null
             cancelLoadingJob()
         }.launchIn(this)
@@ -152,7 +111,7 @@ class Web3NameIdentifierProvider(
             }
     }
 
-    fun ResourceManager.addressWithDescription(w3nAccount: Web3NameAccount): String {
+    private fun ResourceManager.addressWithDescription(w3nAccount: Web3NameAccount): String {
         val description = w3nAccount.description
         return if (description != null) {
             return getString(R.string.web3names_address_with_description, w3nAccount.address.toShortAddressFormat(), description)
